+++ lib/rubocop/rspec/language/each_selector.rb
@@ -1,0 +1,7 @@
+# Monkey-patch SelectorSet to allow enumeration of selectors.
+
+class RuboCop::RSpec::Language::SelectorSet
+  def each(&blk)
+    selectors.each(&blk)
+  end
+end